Clean up code and files
This commit adds multiple help files, explaining how to do various things,
moves files from the top level directory into the package directory, styles
the code according to the defaults of styler, adds more information to the
DESCRIPTION file, creates a .gitignore, and gets started with tests.

It should have probably been multiple smaller commits, but this is what it
ended up as. There still needs to be more work done to automate things and
eventually set up github actions to do this work for us. In the meantime there
may be many commits that are simply "Style with styler".